# src/docs/swagger.yaml
basePath: /
definitions:
  app.VersionInfo:
    properties:
      build:
        type: string
      commitHash:
        type: string
      timestamp:
        type: string
      version:
        type: string
    type: object
  models.APIKey:
    properties:
      description:
        type: string
      key:
        type: string
    type: object
  models.Talent:
    properties:
      description:
        type: string
      id:
        type: string
      maxLvl:
        type: integer
      name:
        type: string
      test:
        type: string
      version:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Talent API with API Key auth
  title: Talent API
  version: "1.0"
paths:
  /admin/keys:
    post:
      consumes:
      - application/json
      parameters:
      - description: API Key description
        in: body
        name: apikey
        schema:
          $ref: '#/definitions/models.APIKey'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.APIKey'
      security:
      - ApiKeyAuth: []
      summary: Create API key
      tags:
      - admin
  /admin/talents:
    post:
      consumes:
      - application/json
      parameters:
      - description: Talent
        in: body
        name: talent
        required: true
        schema:
          $ref: '#/definitions/models.Talent'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Talent'
      security:
      - ApiKeyAuth: []
      summary: Create talent
      tags:
      - admin
  /admin/talents/{id}:
    delete:
      parameters:
      - description: Talent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete talent
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Talent ID
        in: path
        name: id
        required: true
        type: string
      - description: Talent
        in: body
        name: talent
        required: true
        schema:
          $ref: '#/definitions/models.Talent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Talent'
      security:
      - ApiKeyAuth: []
      summary: Update talent
      tags:
      - admin
  /api/talents:
    get:
      consumes:
      - application/json
      description: get talents, optional search
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Talent'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List talents
      tags:
      - talents
  /version:
    get:
      description: Get the version of the server
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.VersionInfo'
      summary: Get the version
      tags:
      - version
securityDefinitions:
  ApiKeyAuth:
    description: Enter your API key (admin or generated)
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"